Fix the internal camera transformation
That was used in order to have a front view of the
scene when the rotation on SDF was <0 0 0>. However,
when no rotations are informed the camera frame have
to match with the body frame, ie, z axis points up and
so the camera has to lock toward the sky. So the current
transformation do it.