fix QApplication intialization and deletion
QApplication requires that the data referred to by argc and argv must
stay valid for the entire lifetime of the QApplication object. In
addition, argc must be greater than zero and argv must contain at least
one valid character string.

Besides this it should not be responsability of the class to spawn the
QApplication, that must be a singleton. Rock already has support to tag
if a component uses qt and this will be utilized in future
implementations. Backward compatibility is assured if someone tries to
run vizkit3dworld withou spawning the application beforehand.